+++ mobile/android/android-components/components/lib/crash/src/main/java/mozilla/components/lib/crash/notification/CrashNotification.kt
@@ -34,13 +34,14 @@ internal class CrashNotification(
     fun show() {
         val pendingIntent = PendingIntent.getActivity(
             context, SharedIdsHelper.getNextIdForTag(context, PENDING_INTENT_TAG),
-            CrashPrompt.createIntent(context, crash), 0
+            CrashPrompt.createIntent(context, crash), getNotificationFlag()
         )
 
         val reportPendingIntent = SendCrashReportService
             .createReportIntent(context, crash, NOTIFICATION_TAG, NOTIFICATION_ID)
             .asForegroundServicePendingIntent(
-                context, SharedIdsHelper.getNextIdForTag(context, PENDING_INTENT_TAG)
+                context, SharedIdsHelper.getNextIdForTag(context, PENDING_INTENT_TAG),
+                getNotificationFlag()
             )
 
         val channel = ensureChannelExists(context)
@@ -101,5 +102,11 @@ internal class CrashNotification(
 
             return NOTIFICATION_CHANNEL_ID
         }
+    }
+
+    private fun getNotificationFlag() = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
+        PendingIntent.FLAG_IMMUTABLE
+    } else {
+        0
     }
 }